tests

+++ src/main/java/org/tinkoff/notifications/controller/AuthController.java
@@ -1,9 +1,8 @@
 package org.tinkoff.notifications.controller;
 
-
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.JdbcUserDetailsManager;
 import org.springframework.web.bind.annotation.*;
@@ -14,13 +13,15 @@ import org.tinkoff.notifications.service.EmployeeService;
 import javax.validation.Valid;
 import java.util.Set;
 
+import static org.tinkoff.notifications.constraint.ApplicationError.ACCESS_DENIED;
+
 @RestController
 @RequestMapping("/auth")
 public class AuthController {
 
-    private EmployeeService employeeService;
-    private JdbcUserDetailsManager userDetailsManager;
-    private PasswordEncoder passwordEncoder;
+    private final EmployeeService employeeService;
+    private final JdbcUserDetailsManager userDetailsManager;
+    private final PasswordEncoder passwordEncoder;
 
     public AuthController(EmployeeService employeeService, JdbcUserDetailsManager userDetailsManager, PasswordEncoder passwordEncoder) {
         this.employeeService = employeeService;
@@ -42,7 +43,13 @@ public class AuthController {
     }
 
     @DeleteMapping("/delete")
-    public void delete(@RequestParam("username") String username) {
+    public void delete(@RequestParam("username") String username, Authentication authentication) {
+        if (!authentication.getName().equals(username)
+                && !authentication
+                        .getAuthorities()
+                        .contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
+            throw ACCESS_DENIED.exception("");
+        }
         userDetailsManager.deleteUser(username);
     }
 }
